requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
setuptools==80.9.0
Werkzeug==3.1.3
wheel==0.45.1
pyodbc==5.2.0
python-dotenv==1.0.0
